requirements.txt
torch
torchvision
lightning
wandb
scikit-learn
pillow
albumentationsx
